requirements.txt
openai-whisper>=20250625
Flask>=3.1.2
openai>=2.6.0
python-docx>=1.2.0
reportlab>=4.4.4
flask-cors>=6.0.1
pytest>=8.3.3
